requirements.txt
g4f==0.1.6.6
scikit-learn==1.3.0
streamlit-option-menu==0.3.6
streamlit==1.27.2
google-generativeai==0.2.1